角色管理功能实现及用户管理页面与功能实现

+++ hotel-pojo/src/main/java/com/hotel/pojo/entity/UserRole.java
@@ -14,10 +14,10 @@ public class UserRole {
     /**
      * 用户id
      */
-    private Long userId;
+    private Integer userId;
 
     /**
      * 角色id
      */
-    private Long roleId;
+    private Integer roleId;
 }
